Move assignPanelId to headerProcessor
