+++ packages/core/src/html/ComponentProcessor.js
@@ -1,6 +1,5 @@
 const cheerio = require('cheerio');
 
-const { findHeaderElement } = require('./headerProcessor');
 const { getVslotShorthandName } = require('./vueSlotSyntaxProcessor');
 const { warnConflictingAttributes, warnDeprecatedAttributes } = require('./warnings');
 
@@ -21,33 +20,6 @@ class ComponentProcessor {
   processPanelAttributes(node) {
     this.markdownProcessor.processAttributeWithoutOverride(node, 'alt', false, '_alt');
     this.markdownProcessor.processAttributeWithoutOverride(node, 'header', false);
-  }
-
-  /**
-   * Assigns an id to the root element of a panel component using the heading specified in the
-   * panel's header slot or attribute (if any), with the header slot having priority if present.
-   * This is to ensure anchors still work when panels are in their minimized form.
-   * @param node The root panel element
-   */
-  static assignPanelId(node) {
-    const slotChildren = node.children
-      && node.children.filter(child => getVslotShorthandName(child) !== '');
-
-    const headerSlot = slotChildren.find(child => getVslotShorthandName(child) === 'header');
-
-    if (headerSlot) {
-      const header = findHeaderElement(headerSlot);
-      if (!header) {
-        return;
-      }
-
-      if (!header.attribs || !_.has(header.attribs, 'id')) {
-        throw new Error('Found a panel heading without an assigned id.\n'
-          + 'Please report this to the MarkBind developers. Thank you!');
-      }
-
-      node.attribs.id = header.attribs.id;
-    }
   }
 
   /*
